refactor: consolidate players page
