--- src/app/players/PlayersPageServer.tsx
@@ -1,5 +1,0 @@
-import PlayersPageClient from './PlayersPageClient';
-
-export default function PlayersPageServer() {
-  return <PlayersPageClient />;
-}
